moved NodeMeta into separate file
