+++ core/record/node.js
@@ -8,6 +8,8 @@ import * as Validation from '@core/validation/validation'
 import * as NodeDef from '@core/survey/nodeDef'
 import * as NodeDefLayout from '@core/survey/nodeDefLayout'
 
+import { NodeMeta } from './_node/nodeMeta'
+
 export const keys = {
   id: ObjectUtils.keys.id,
   uuid: ObjectUtils.keys.uuid,
@@ -17,20 +19,13 @@ export const keys = {
   recordUuid: 'recordUuid',
   nodeDefUuid: ObjectUtils.keys.nodeDefUuid,
   value: 'value',
-  meta: 'meta',
+  meta: NodeMeta.keys.meta,
   placeholder: 'placeholder',
 
   created: 'created',
   updated: 'updated',
   deleted: 'deleted',
   dirty: 'dirty', // Modified by the user but not persisted yet
-}
-
-export const metaKeys = {
-  hierarchy: 'h', // Ancestor nodes uuids hierarchy
-  childApplicability: 'childApplicability', // Applicability by child def uuid
-  defaultValue: 'defaultValue', // True if default value has been applied, false if the value is user defined
-  hierarchyCode: 'hCode', // Hierarchy of code attribute ancestors (according to the parent code defs specified)
 }
 
 export const valuePropsCode = {
@@ -135,18 +130,10 @@ export const { getValidation } = Validation
 
 // ===== READ metadata
 
-export const getMeta = R.propOr({}, keys.meta)
+export const { metaKeys, getMeta, isChildApplicable, isDefaultValueApplied, getHierarchy, getHierarchyCode } = NodeMeta
 
-export const isChildApplicable = (childDefUuid) =>
-  R.pathOr(true, [keys.meta, metaKeys.childApplicability, childDefUuid])
-export const isDefaultValueApplied = R.pathOr(false, [keys.meta, metaKeys.defaultValue])
-
-export const getHierarchy = R.pathOr([], [keys.meta, metaKeys.hierarchy])
-
+// Hierarchy
 export const isDescendantOf = (ancestor) => (node) => R.includes(getUuid(ancestor), getHierarchy(node))
-
-// Code metadata
-export const getHierarchyCode = R.pathOr([], [keys.meta, metaKeys.hierarchyCode])
 
 //
 // ======
@@ -180,53 +167,12 @@ export const newNodePlaceholder = (nodeDef, parentNode, value = null) => ({
 export const assocValue = R.assoc(keys.value)
 export const { assocValidation } = Validation
 
-export const assocMeta = R.assoc(keys.meta)
-
-export const mergeMeta = (meta) => (node) => {
-  const metaOld = getMeta(node)
-  const metaUpdated = R.mergeLeft(meta)(metaOld)
-  return assocMeta(metaUpdated)(node)
-}
-
-export const assocIsDefaultValueApplied = (value) => (node) => {
-  const metaKey = metaKeys.defaultValue
-  const metaOld = getMeta(node)
-  const metaUpdated = { ...metaOld }
-  if (value) {
-    metaUpdated[metaKey] = true
-  } else {
-    // default value is false by default
-    delete metaUpdated[metaKey]
-  }
-  return assocMeta(metaUpdated)(node)
-}
-
-export const assocChildApplicability =
-  ({ nodeDefUuid, applicable }) =>
-  (node) => {
-    const metaKey = metaKeys.childApplicability
-    const metaOld = getMeta(node)
-    const childApplicabilityOld = metaOld[metaKey]
-    const childApplicabilityUpdated = { ...childApplicabilityOld }
-    if (applicable) {
-      // applicable is true by default, remove it from meta object
-      delete childApplicabilityUpdated[nodeDefUuid]
-    } else {
-      childApplicabilityUpdated[nodeDefUuid] = false
-    }
-    const metaUpdated = { ...metaOld }
-    if (R.isEmpty(childApplicabilityUpdated)) {
-      delete metaUpdated[metaKey]
-    } else {
-      metaUpdated[metaKey] = childApplicabilityUpdated
-    }
-    return assocMeta(metaUpdated)(node)
-  }
+export const { assocMeta, mergeMeta, assocChildApplicability, assocIsDefaultValueApplied } = NodeMeta
 
 export const assocCreated = R.assoc(keys.created)
 export const assocDeleted = R.assoc(keys.deleted)
 export const assocUpdated = R.assoc(keys.updated)
- 
+
 //
 // ======
 // UTILS
